Fixed Glitch
Changed starting prevX prevY and scaling factor of Yaw and Pitch to make
sure the view makes sense. a person can turn 360 degrees left and right
(yaw) but can only see up and down 180 degrees (pitch). It also fixes
the glitch where it turns upsidedown for pitch because it went past the
180 degree mark.